+++ ohsome_quality_analyst/api/request_models.py
@@ -1,9 +1,8 @@
 import json
 
 import geojson
-import pydantic
 from geojson import FeatureCollection
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict, Field, field_validator
 
 from ohsome_quality_analyst.topics.definitions import TopicEnum
 from ohsome_quality_analyst.topics.models import TopicData
@@ -11,10 +10,40 @@ from ohsome_quality_analyst.utils.helper import snake_to_lower_camel
 from ohsome_quality_analyst.utils.validators import validate_geojson
 
 
-class BaseBpolys(BaseModel):
-    bpolys: FeatureCollection
+class BaseConfig(BaseModel):
+    model_config = ConfigDict(
+        alias_generator=snake_to_lower_camel,
+        populate_by_name=True,
+        frozen=True,
+        extra="forbid",
+    )
 
-    @pydantic.validator("bpolys")
+
+class BaseBpolys(BaseConfig):
+    bpolys: dict = Field(
+        {
+            "type": "FeatureCollection",
+            "features": [
+                {
+                    "type": "Feature",
+                    "geometry": {
+                        "type": "Polygon",
+                        "coordinates": [
+                            [
+                                [8.674092292785645, 49.40427147224242],
+                                [8.695850372314453, 49.40427147224242],
+                                [8.695850372314453, 49.415552187316095],
+                                [8.674092292785645, 49.415552187316095],
+                                [8.674092292785645, 49.40427147224242],
+                            ]
+                        ],
+                    },
+                },
+            ],
+        }
+    )
+
+    @field_validator("bpolys")
     @classmethod
     def validate_bpolys(cls, value) -> FeatureCollection:
         obj = geojson.loads(json.dumps(value))
@@ -25,44 +54,13 @@ class BaseBpolys(BaseModel):
 
 
 class IndicatorRequest(BaseBpolys):
-    topic_key: TopicEnum = pydantic.Field(
+    topic_key: TopicEnum = Field(
         ...,
         title="Topic Key",
         alias="topic",
-        example="building-count",
     )
+    include_figure: bool = True
     include_data: bool = False
-
-    class Config:
-        """Pydantic config class."""
-
-        alias_generator = snake_to_lower_camel
-        # Allow population by field name not just by alias name
-        allow_population_by_field_name = True
-        allow_mutation = False
-        extra = "forbid"
-        schema_extra = {
-            "examples": [
-                {
-                    "topic": "building-count",
-                    "bpolys": {
-                        "type": "Feature",
-                        "geometry": {
-                            "type": "Polygon",
-                            "coordinates": [
-                                [
-                                    [8.674092292785645, 49.40427147224242],
-                                    [8.695850372314453, 49.40427147224242],
-                                    [8.695850372314453, 49.415552187316095],
-                                    [8.674092292785645, 49.415552187316095],
-                                    [8.674092292785645, 49.40427147224242],
-                                ]
-                            ],
-                        },
-                    },
-                }
-            ]
-        }
 
 
 class IndicatorDataRequest(BaseBpolys):
@@ -71,27 +69,10 @@ class IndicatorDataRequest(BaseBpolys):
     The Topic consists of name, description and data.
     """
 
+    topic: TopicData = Field(..., title="Topic", alias="topic")
+    include_figure: bool = True
     include_data: bool = False
-    topic: TopicData = pydantic.Field(..., title="Topic", alias="topic")
-
-    class Config:
-        """Pydantic config class."""
-
-        alias_generator = snake_to_lower_camel
-        # Allow population by field name not just by alias name
-        allow_population_by_field_name = True
-        allow_mutation = False
-        extra = "forbid"
 
 
 class ReportRequest(BaseBpolys):
     include_data: bool = False
-
-    class Config:
-        """Pydantic config class."""
-
-        alias_generator = snake_to_lower_camel
-        # Allow population by field name not just by alias name
-        allow_population_by_field_name = True
-        allow_mutation = False
-        extra = "forbid"