+++ tests/integrationtests/indicators/test_building_completeness.py
@@ -130,7 +130,6 @@ class TestCalculationFigure:
         indicator.create_figure()
         assert isinstance(indicator.result.figure, dict)
         pgo.Figure(indicator.result.figure)  # test for valid Plotly figure
-        assert indicator.result.svg is not None
 
 
 class TestGetData: