+++ tests/integrationtests/indicators/test_mapping_saturation.py
@@ -120,7 +120,6 @@ class TestFigure:
         indicator.create_figure()
         assert isinstance(indicator.result.figure, dict)
         pgo.Figure(indicator.result.figure)  # test for valid Plotly figure
-        assert indicator.result.svg is not None
 
     @pytest.mark.skip(reason="Only for manual testing.")  # comment for manual test
     def test_create_figure_with_high_asymptote_manual(self, indicator):
@@ -163,7 +162,7 @@ class TestFigure:
 @oqt_vcr.use_cassette
 def test_immutable_attribute(
     topic_building_count,
-    feature_collection_germany_heidelberg_bahnstadt_bergheim,
+    feature_collection_heidelberg_bahnstadt_bergheim_weststadt,
 ):
     """Test changes of attribute values when multiple indicators are created.
 
@@ -174,7 +173,9 @@ def test_immutable_attribute(
     """
     indicators = []
     fitted_values = []
-    for feature in feature_collection_germany_heidelberg_bahnstadt_bergheim["features"]:
+    for feature in feature_collection_heidelberg_bahnstadt_bergheim_weststadt[
+        "features"
+    ]:
         indicator = MappingSaturation(topic_building_count, feature)
         asyncio.run(indicator.preprocess())
         indicator.calculate()