+++ tests/unittests/test_report_minimal.py
@@ -15,7 +15,6 @@ class TestReportMinimal(unittest.TestCase):
         for _ in report.indicator_topic:
             indicator = Mock()
             indicator.result = Mock()
-            indicator.result.html = "foo"
             indicator.result.class_ = 1
             report.indicators.append(indicator)
 
@@ -23,7 +22,6 @@ class TestReportMinimal(unittest.TestCase):
 
         self.assertIsNotNone(report.result.label)
         self.assertIsNotNone(report.result.description)
-        self.assertIsNotNone(report.result.html)
         self.assertEqual(report.result.class_, 1)